feat: Refactor font size and line height variables into separate files
